Reference

All docstrings.

AstrophysicalCalculations.AstrophysicalCalculations — Module

Equations common to introductory astrophysics coursework. All calculations are directly referenced from the second edition of An Introduction to Modern Astrophysics by Carroll and Ostlie.

Extended Help

Quantities

This package uses DynamicQuantities.jl for all unit-handling. For some calculations, constants from the DynamicQuantities.Cosntants module are used, i.e. parsecs pc. See the DynamicQuantities.jl documentation for more information about converting to and from quantity types, and for interoperability instructions with other unit-handling packages within Julia (namely, Unitful.jl).

References

Reference information for the primary text, An Introduction to Modern Astrophysics, is shown via BibTeX below.

@book{BOB,
    title={An Introduction to Modern Astrophysics},
    author={Carroll, Bradley W and Ostlie, Dale A},
    year={2017},
    publisher={Cambridge University Press}
}

Imports

Base

Core

DocStringExtensions

DynamicQuantities

Reexport

Exports

L☉

M☉

absolute_magnitude

absolute_magnitude_difference

blackbody_frequency

blackbody_luminosity

blackbody_temperature

blackbody_wavelength

doppler_shift

effective_luminosity

effective_temperature

flux_angle

flux_distance

flux_ratio

k

length_contraction

lorentz_factor

mean_poynting

momentum_magnitude

parallax_distance

poynting

radiant_flux

redshift_parameter

relativistic_kinetic_energy

time_dilation

total_relativistic_energy

wave_speed

σ

AstrophysicalCalculations.L☉ — Constant

The luminosity of the Sun.

Extended Help

Caveats

The use of this value assumes the inverse square law for light.

References

(Carrol & Ostlie 62)

AstrophysicalCalculations.M☉ — Constant

The absolute magnitude of the Sun.

Extended Help

Caveats

The use of this value assumes the inverse square law for light.

References

(Carrol & Ostlie 62)

AstrophysicalCalculations.k — Constant

The Boltzmann constant k.

Extended Help

Caveats

This constant is used in Planck's function.

References

(Carrol & Ostlie 72)

AstrophysicalCalculations.σ — Constant

The Stefan-Boltzmann constant σ.

Extended Help

Caveats

This constant is used in the Stefan-Boltzmann equation, and is not physically accurate due to its assumption of a continuous spectrum of light.

References

(Carrol & Ostlie 70)

AstrophysicalCalculations.absolute_magnitude — Method

absolute_magnitude(L, Mᵣ, Lᵣ)

The absolute magnitude of a star, given its luminosity L, using another star's absolute magnitude Mᵣ and luminosity Lᵣ as a reference.

Extended Help

Caveats

This calculation assumes the inverse square law for light, and uses some other star's absolute magnitude and luminosity as a reference.

References

(Carrol & Ostlie 62)

AstrophysicalCalculations.absolute_magnitude — Method

absolute_magnitude(L)

The absolute magnitude of a star, given its luminosity L, using our Sun's absolute magnitude and luminosity as a reference.

Extended Help

Caveats

This calculation assumes the inverse square law for light, and uses our Sun's absolute magnitude and luminosity as a reference.

References

(Carrol & Ostlie 62)

AstrophysicalCalculations.absolute_magnitude_difference — Method

absolute_magnitude_difference(F₁, F₂)

The difference in absolute magnitudes mᵢ for two stars. The absolute magnitude M is defined as the apparent magnitude of the star if it were located 10 parsecs (pc) away.

Extended Help

Caveats

This calculation assumes the inverse square law for light.

References

(Carrol & Ostlie 61)

AstrophysicalCalculations.absorption_pressure — Method

absorption_pressure(S₀, A, θ)

The radiation pressure exerted on a surface which absorbs light, where θ is the angle of incidence of the light on the surface, and A is the area of the surface.

Extended Help

Caveats

This equation is valid for surfaces which completely absorb light.

References

(Carrol & Ostlie 67)

AstrophysicalCalculations.blackbody_frequency — Method

blackbody_frequency(ν, T)

The blackbody spectra at some frequency ν and temperature T. This re-formulation of planck's function builds on top of the Relleigh-Jean equation.

Extended Help

Caveats

This equation assumes the inverse square law of light.

References

(Carrol & Ostlie 73)

AstrophysicalCalculations.blackbody_luminosity — Method

blackbody_luminosity(A, T)

The luminosity of some surface with area A, and blackbody temperature T in Kelvin.

Extended Help

Caveats

This equation assumes a blackbody temperature provided in Kelvin.

References

(Carrol & Ostlie 70)

AstrophysicalCalculations.blackbody_temperature — Method

blackbody_temperature(λₘ)

The blackbody temperature T, given the maximum wavelength of emitted light, according to Wien's displacement law.

Extended Help

Caveats

This equation is valid for continuous spectrum of light; therefore, it is not physically realistic! The spectrum of light is discrete, not continuous.

References

(Carrol & Ostlie 69)

AstrophysicalCalculations.blackbody_wavelength — Method

blackbody_wavelength(λ, T)

The blackbody spectra at some wavelength λ and temperature T. This equation was derived by Planck, and builds on top of the Relleigh-Jean equation.

Extended Help

Caveats

This equation assumes the inverse square law of light.

References

(Carrol & Ostlie 73)

AstrophysicalCalculations.bolometric_correction — Method

bolometric_correction(Mᵦ, Mᵥ)

The bolometric correction BC, which is defined as the difference between a star's bolometric magnitude Mᵦ and its visual magnitude Mᵥ.

Extended Help

Caveats

This equation assumes the inverse square law of light.

References

(Carrol & Ostlie 76)

AstrophysicalCalculations.constructive_interference — Method

constructive_interference(n, λ)

The positions d * sin(θ) of constructive interference for the double-slit experiment.

Extended Help

Caveats

This calculation assumes n multiples of λ, where n is between 0 and inf, inclusive.

References

(Carrol & Ostlie 65)

AstrophysicalCalculations.destructive_interference — Method

destructive_interference(n, λ)

The positions d * sin(θ) of destructive interference for the double-slit experiment.

Extended Help

Caveats

This calculation assumes n multiples of λ, where n is between 1 and inf, inclusive.

References

(Carrol & Ostlie 65)

AstrophysicalCalculations.distance_modulus — Method

distance_modulus(d)

The difference between apparent and absolute magnitudes, m - M, of a star, given the distance d.

Extended Help

Caveats

This calculation assumes the inverse square law for light.

References

(Carrol & Ostlie 62)

AstrophysicalCalculations.doppler_shift — Method

doppler_shift(ν, vᵣ, u)

The relativistic doppler shift observed from an object moving with speed u away from the light, rest frequency ν, and radial velocity vᵣ.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 98)

AstrophysicalCalculations.effective_luminosity — Method

effective_luminosity(R, Tₑ)

The luminosity of some spherical star with radius R, and effective temperature Tₑ.

Extended Help

Caveats

This equation assumes the inverse square law of light.

References

(Carrol & Ostlie 70)

AstrophysicalCalculations.effective_temperature — Method

effective_temperature(F)

The effective temperature of some spherical star with radiant flux F.

Extended Help

Caveats

This equation assumes the inverse square law of light, and the Stefan-Boltzmann constant.

References

(Carrol & Ostlie 70)

AstrophysicalCalculations.flux_angle — Method

flux_angle(γ)

The angle of radiant flux for a moving emitter given the Lorentz factor γ; this phenomena is known as the relativistic headlight effect.

Extended Help

Caveats

This calculation is valid for velocities very close to the speed of light, and within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 101)

AstrophysicalCalculations.flux_distance — Method

flux_distance(m, M)

The distance to a star in parsecs (pc), given the apparent and absolute magnitudes.

Extended Help

Caveats

This calculation assumes the inverse square law for light.

References

(Carrol & Ostlie 62)

AstrophysicalCalculations.flux_ratio — Method

flux_ratio(m₁, m₂)

The flux ratio for two stars, which is defined as the apparent magnitude of the star if it were located 10 parsecs (pc) away, given two apparent magnitudes m₁ and m₂.

Extended Help

Caveats

This calculation assumes the inverse square law for light.

References

(Carrol & Ostlie 61)

AstrophysicalCalculations.length_contraction — Method

length_contraction(Lᵣ, u)

The relativistic length scaling used within Lorentz Transformations, given reference velocity u. This calculation returns the length of an object in a moving reference frame, given the length of the same object in an at-rest frame Lᵣ, and the velocity u.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 95)

AstrophysicalCalculations.lorentz_factor — Method

lorentz_factor(u)

The relativistic factor γ that is within Lorentz Transformations, given reference velocity u.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 91)

AstrophysicalCalculations.mean_poynting — Method

mean_poynting(E₀, B₀)

The time-average of the Poynting vector S, which describes the rate that a light wave carries its energy, given electric and magnetic fields E and B. This vector points in the direction of the electromagnetic wave's propagation.

Extended Help

Caveats

Electric and magnetic field strengths vary with time, so the time-averaged value of this vector is what is often physically important.

References

(Carrol & Ostlie 67)

AstrophysicalCalculations.momentum — Method

momentum(m, v)

The momentum of a particle moving at relativistic speeds, given mass m and velocity v.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 102)

AstrophysicalCalculations.momentum_magnitude — Method

momentum_magnitude(m, E)

The magnitude a particle's momentum p, given mass m and total energy E.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 103)

AstrophysicalCalculations.parallax_distance — Method

parallax_distance(p, B)

The distance d to a cellestial object with parallax angle p and baseline distance B.

Extended Help

Caveats

The units of this calculation are equivalent to the units of B.

References

(Carrol & Ostlie 58)

AstrophysicalCalculations.parallax_distance — Method

parallax_distance(p′′)

The distance d in parsecs to a cellestial object with parallax angle p′′. One parsec is equivalent to 2.06264806e5 Astronomical Units (AU).

Extended Help

Caveats

This calculation requires the small-angle approximation.

References

(Carrol & Ostlie 58)

AstrophysicalCalculations.poynting — Method

poynting(E, B)

The Poynting vector S, which describes the rate that a light wave carries its energy, given electric and magnetic fields E and B. This vector points in the direction of the electromagnetic wave's propagation.

Extended Help

Caveats

Electric and magnetic field strengths vary with time, so the time-averaged value of this vector is what is often physically important.

References

(Carrol & Ostlie 65)

AstrophysicalCalculations.radiant_flux — Method

radiant_flux(L, r)

The radiant flux F captured by a spherical shell with radius r, emitted by a star with luminosity L, assuming no light is absorbed between the star's exterior and the shell at radius r.

Extended Help

Caveats

This calculation assumes the inverse square law for light, and assumes no light is absorbed between the star's exterior and the shell at radius r.

References

(Carrol & Ostlie 61)

AstrophysicalCalculations.redshift_parameter — Method

redshift_parameter(Δλ, λᵣ)

The redshift parameter z, which describes the change in wavelength due to redshifting with change in wavelength Δλ and rest wavelength λᵣ, or radial velocity vᵣ.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 99)

AstrophysicalCalculations.reflection_pressure — Method

reflection_pressure(S₀, A, θ)

The radiation pressure exerted on a surface which reflects light, where θ is the angle of incidence of the light on the surface, and A is the area of the surface.

Extended Help

Caveats

This equation is valid for surfaces which completely reflect light.

References

(Carrol & Ostlie 67)

AstrophysicalCalculations.relativistic_kinetic_energy — Method

relativistic_kinetic_energy(m, v)

The kinetic energy K of a particle moving at relativistic speeds, given mass m and velocity v.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 103)

AstrophysicalCalculations.time_dilation — Method

time_dilation(Δtᵣ, u)

The relativistic time scaling used within Lorentz Transformations, given reference velocity u. This calculation returns the change in time of a moving reference frame, given the change in time in an at-rest frame Δtᵣ, and the velocity u.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 93)

AstrophysicalCalculations.total_relativistic_energy — Method

total_relativistic_energy(m, v)

The total energy E of a particle moving at relativistic speeds, given mass m and velocity v.

Extended Help

Caveats

This calculation is valid within the bounds of special relativity: that is, inertial reference frames.

References

(Carrol & Ostlie 103)

AstrophysicalCalculations.wave_speed — Method

wave_speed(λ, ν)

The speed of a wave, given its wavelength λ and frequency ν.

Extended Help

Caveats

This calculation is not physically realistic for light waves; the speed of light is a fundamental constant of nature: 2.99792458e8 m s⁻¹.

References

(Carrol & Ostlie 63)